lesson 36

+++ src/lesson35/controller/HotelController.java
@@ -4,25 +4,30 @@ import lesson35.model.Hotel;
 import lesson35.service.HotelService;
 
 public class HotelController {
+    private UserController userController = new UserController();
     private HotelService hotelService = new HotelService();
 
-    public void addHotel(Hotel hotel){
+    public void addHotel(Hotel hotel) throws Exception {
+        userController.validateLoginAdmin();
         hotelService.addHotel(hotel);
     }
 
-    public Hotel findHotelByName(String name){
+    public Hotel findHotelByName(String name) throws Exception {
+        userController.validateLogin();
         return hotelService.findHotelByName(name);
     }
 
-    public Hotel findHotelByCity(String city){
+    public Hotel findHotelByCity(String city) throws Exception {
+        userController.validateLogin();
         return hotelService.findHotelByCity(city);
     }
 
-    public Hotel findHotelById(long id){
+    public Hotel findHotelById(long id) {
         return hotelService.findHotelById(id);
     }
 
-    public void deleteHotel(long id){
+    public void deleteHotel(long id) throws Exception {
+        userController.validateLoginAdmin();
         hotelService.deleteHotel(id);
     }
 }
