+++ src/lesson35/controller/RoomController.java
@@ -7,21 +7,25 @@ import lesson35.service.RoomService;
 import java.util.ArrayList;
 
 public class RoomController {
+    private UserController userController = new UserController();
     private RoomService roomService = new RoomService();
 
-    public void addRoom(Room room){
+    public void addRoom(Room room) throws Exception {
+        userController.validateLoginAdmin();
         roomService.addRoom(room);
     }
 
-    public Room findRoomById(long id){
+    public Room findRoomById(long id) {
         return roomService.findRoomById(id);
     }
 
-    public void deleteRoom(long id){
+    public void deleteRoom(long id) throws Exception {
+        userController.validateLoginAdmin();
         roomService.deleteRoom(id);
     }
 
-    public ArrayList<Room> findRooms(Filter filter){
+    public ArrayList<Room> findRooms(Filter filter) throws Exception {
+        userController.validateLogin();
         return roomService.findRooms(filter);
     }
 }
